# quickselect: implement several optimizations